Extends ArtifactBaselines and centralizes conversion logic
