+++ src/main/kotlin/io/dotanuki/arw/features/baseline/GenerateCommand.kt
@@ -5,10 +5,9 @@ import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import io.dotanuki.arw.core.domain.errors.ErrorAware
-import io.dotanuki.arw.core.domain.models.AnalysedArtifact
-import io.dotanuki.arw.core.domain.models.AndroidComponentType
 import io.dotanuki.arw.core.infrastructure.android.AndroidArtifactAnalyser
 import io.dotanuki.arw.core.infrastructure.cli.ErrorReporter
+import io.dotanuki.arw.features.common.ArtifactBaseline
 
 context (BaselineContext)
 class GenerateCommand : CliktCommand(
@@ -25,20 +24,7 @@ class GenerateCommand : CliktCommand(
         val analysed = AndroidArtifactAnalyser.overview(target)
         val outputFile = "${analysed.applicationId}.toml"
 
-        val baseline = with(analysed) {
-            ArtifactBaseline(
-                androidPermissions,
-                androidFeatures,
-                registeredInstances(AndroidComponentType.ACTIVITY),
-                registeredInstances(AndroidComponentType.SERVICE),
-                registeredInstances(AndroidComponentType.RECEIVER),
-                registeredInstances(AndroidComponentType.PROVIDER)
-            )
-        }
-
+        val baseline = ArtifactBaseline.from(analysed)
         BaselineWriter.write(baseline, outputFile)
     }
-
-    private fun AnalysedArtifact.registeredInstances(componentType: AndroidComponentType) =
-        androidComponents.filter { it.type == componentType }.map { it.name }
 }
