+++ src/main/kotlin/io/dotanuki/arw/features/comparison/DetectedChange.kt
@@ -2,5 +2,5 @@ package io.dotanuki.arw.features.comparison
 
 data class DetectedChange(
     val what: String,
-    val nature: ProposedAction
+    val action: ProposedAction
 )
