+++ src/main/kotlin/io/dotanuki/arw/features/overview/ArtifactOverview.kt
@@ -7,7 +7,6 @@ data class ArtifactOverview(
     val applicationId: String,
     val minSdk: Int,
     val targetSdk: Int,
-    val debuggable: Boolean,
     val totalUsedFeatures: Int,
     val totalPermissions: Int,
     val dangerousPermissions: Boolean,
